+++ core/CHANGELOG.md
@@ -5,6 +5,8 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 
 ## 1.0.22
 
++ [#1962](https://github.com/luyadev/luya/issues/1962) Fix problem with first stack trace informations not containing line and file informations.
+
 ## 1.0.21 (9. September 2019)
 
 + [#1956](https://github.com/luyadev/luya/issues/1956) Ensure a not existing package installer config value `installer.php` won't throw an exception.